chore: move messages into their own file
